# docs/openapi.yml
components:
  schemas:
...
    EsiScope:
      description: Enumeration of the ESI scopes
      enum:
      - esi-alliances.read_contacts.v1
      - esi-assets.read_assets.v1
      - esi-assets.read_corporation_assets.v1
      - esi-bookmarks.read_character_bookmarks.v1
      - esi-bookmarks.read_corporation_bookmarks.v1
      - esi-calendar.read_calendar_events.v1
      - esi-calendar.respond_calendar_events.v1
      - esi-characters.read_agents_research.v1
      - esi-characters.read_blueprints.v1
      - esi-characters.read_chat_channels.v1
      - esi-characters.read_contacts.v1
      - esi-characters.read_corporation_roles.v1
      - esi-characters.read_fatigue.v1
      - esi-characters.read_fw_stats.v1
      - esi-characters.read_loyalty.v1
      - esi-characters.read_medals.v1
      - esi-characters.read_notifications.v1
      - esi-characters.read_opportunities.v1
      - esi-characters.read_standings.v1
      - esi-characters.read_titles.v1
      - esi-characters.write_contacts.v1
      - esi-characterstats.read.v1
      - esi-clones.read_clones.v1
      - esi-clones.read_implants.v1
      - esi-contracts.read_character_contracts.v1
      - esi-contracts.read_corporation_contracts.v1
      - esi-corporations.read_blueprints.v1
      - esi-corporations.read_contacts.v1
      - esi-corporations.read_container_logs.v1
      - esi-corporations.read_corporation_membership.v1
      - esi-corporations.read_divisions.v1
      - esi-corporations.read_facilities.v1
      - esi-corporations.read_fw_stats.v1
      - esi-corporations.read_medals.v1
      - esi-corporations.read_standings.v1
      - esi-corporations.read_starbases.v1
      - esi-corporations.read_structures.v1
      - esi-corporations.read_titles.v1
      - esi-corporations.track_members.v1
      - esi-fittings.read_fittings.v1
      - esi-fittings.write_fittings.v1
      - esi-fleets.read_fleet.v1
      - esi-fleets.write_fleet.v1
      - esi-industry.read_character_jobs.v1
      - esi-industry.read_character_mining.v1
      - esi-industry.read_corporation_jobs.v1
      - esi-industry.read_corporation_mining.v1
      - esi-killmails.read_corporation_killmails.v1
      - esi-killmails.read_killmails.v1
      - esi-location.read_location.v1
      - esi-location.read_online.v1
      - esi-location.read_ship_type.v1
      - esi-mail.organize_mail.v1
      - esi-mail.read_mail.v1
      - esi-mail.send_mail.v1
      - esi-markets.read_character_orders.v1
      - esi-markets.read_corporation_orders.v1
      - esi-markets.structure_markets.v1
      - esi-planets.manage_planets.v1
      - esi-planets.read_customs_offices.v1
      - esi-search.search_structures.v1
      - esi-skills.read_skillqueue.v1
      - esi-skills.read_skills.v1
      - esi-ui.open_window.v1
      - esi-ui.write_waypoint.v1
      - esi-universe.read_structures.v1
      - esi-wallet.read_character_wallet.v1
      - esi-wallet.read_corporation_wallet.v1
      - esi-wallet.read_corporation_wallets.v1
      - publicData
      title: EsiScope
      type: string
...
    GetCoalitionOut:
      description: Model for `GET /coalition/{coalition_id}` responses.
      properties:
        authorized_to_login:
          title: Authorized To Login
          type: boolean
        coalition_id:
          title: Coalition Id
          type: string
        coalition_name:
          title: Coalition Name
          type: string
        created_on:
          format: date-time
          title: Created On
          type: string
        mandatory_esi_scopes:
          items:
            $ref: '#/components/schemas/EsiScope'
          title: Mandatory Esi Scopes
          type: array
        member_alliances:
          items:
            $ref: '#/components/schemas/sni__api__routers__coalition__GetAllianceShortOut'
          title: Member Alliances
          type: array
        member_corporations:
          items:
            $ref: '#/components/schemas/GetCorporationShortOut'
          title: Member Corporations
          type: array
        ticker:
          title: Ticker
          type: string
        updated_on:
          format: date-time
          title: Updated On
          type: string
      required:
      - coalition_id
      - created_on
      - mandatory_esi_scopes
      - member_alliances
      - member_corporations
      - coalition_name
      - ticker
      - updated_on
      title: GetCoalitionOut
      type: object
    GetCoalitionShortOut:
      description: Model for an element of a `GET /coalition` response.
      properties:
        coalition_id:
          title: Coalition Id
          type: string
        coalition_name:
          title: Coalition Name
          type: string
      required:
      - coalition_id
      - coalition_name
      title: GetCoalitionShortOut
      type: object
...
    GetCorporationShortOut:
      description: Short corporation description
      properties:
        corporation_id:
          title: Corporation Id
          type: integer
        corporation_name:
          title: Corporation Name
          type: string
      required:
      - corporation_id
      - corporation_name
      title: GetCorporationShortOut
      type: object
...
    GetTrackingOut:
      description: Represents a corporation tracking response.
      properties:
        invalid_refresh_token:
          default: []
          items:
            $ref: '#/components/schemas/GetUserShortOut'
          title: Invalid Refresh Token
          type: array
        no_refresh_token:
          default: []
          items:
            $ref: '#/components/schemas/GetUserShortOut'
          title: No Refresh Token
          type: array
        valid_refresh_token:
          default: []
          items:
            $ref: '#/components/schemas/GetUserShortOut'
          title: Valid Refresh Token
          type: array
      title: GetTrackingOut
      type: object
...
    GetUserShortOut:
      description: Model for an element of ``GET /user`` response
      properties:
        character_id:
          title: Character Id
          type: integer
        character_name:
          title: Character Name
          type: string
      required:
      - character_id
      - character_name
      title: GetUserShortOut
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    PostCoalitionIn:
      description: Model for `POST /coalition` responses.
      properties:
        coalition_name:
          title: Coalition Name
          type: string
        ticker:
          title: Ticker
          type: string
      required:
      - coalition_name
      - ticker
      title: PostCoalitionIn
      type: object
...
    PutCoalitionIn:
      description: Model for `PUT /coalition/{coalition_id}` responses.
      properties:
        add_member_alliances:
          items:
            type: integer
          title: Add Member Alliances
          type: array
        add_member_corporations:
          items:
            type: integer
          title: Add Member Corporations
          type: array
        authorized_to_login:
          title: Authorized To Login
          type: boolean
        mandatory_esi_scopes:
          items:
            $ref: '#/components/schemas/EsiScope'
          title: Mandatory Esi Scopes
          type: array
        member_alliances:
          items:
            type: integer
          title: Member Alliances
          type: array
        member_corporations:
          items:
            type: integer
          title: Member Corporations
          type: array
        remove_member_alliances:
          items:
            type: integer
          title: Remove Member Alliances
          type: array
        remove_member_corporations:
          items:
            type: integer
          title: Remove Member Corporations
          type: array
        ticker:
          title: Ticker
          type: string
      title: PutCoalitionIn
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            type: string
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
    sni__api__routers__coalition__GetAllianceShortOut:
      description: Short alliance description
      properties:
        alliance_id:
          title: Alliance Id
          type: integer
        alliance_name:
          title: Alliance Name
          type: string
      required:
      - alliance_id
      - alliance_name
      title: GetAllianceShortOut
      type: object
...
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.0.2
paths:
...
  /coalition:
    get:
      description: Lists all the coalition names. Requires a clearance level of 0
        or more.
      operationId: get_coalition_coalition_get
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/GetCoalitionShortOut'
                title: Response Get Coalition Coalition Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List all coalition names
      tags:
      - Coalition management
    post:
      description: Creates a coalition. Requires a clearance level of 9 or more.
      operationId: post_coalitions_coalition_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCoalitionIn'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCoalitionOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create a coalition
      tags:
      - Coalition management
  /coalition/{coalition_id}:
    delete:
      description: Deletes a coalition. Requires a clearance level of 9 or more.
      operationId: delete_coalition_coalition__coalition_id__delete
      parameters:
      - in: path
        name: coalition_id
        required: true
        schema:
          title: BSON object id
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete a coalition
      tags:
      - Coalition management
    get:
      description: 'Returns details about a given coalition. Requires a clearance
        level of 0 or

        more.'
      operationId: get_coalition_name_coalition__coalition_id__get
      parameters:
      - in: path
        name: coalition_id
        required: true
        schema:
          title: BSON object id
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCoalitionOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get basic informations about a coalition
      tags:
      - Coalition management
    put:
      description: 'Updates a coalition. All fields in the request body are optional.
        The

        `add_member_alliances` and `remove_member_alliances` fields can be used together,
        but the

        `member_alliances` cannot be used in conjunction with `add_member_alliances`
        and

        `remove_member_alliances`. Similarly for `add_member_corporations`,

        `remove_member_corporations`, and `member_corporations`. Requires a

        clearance level of 6 or more.'
      operationId: put_coalition_coalition__coalition_id__put
      parameters:
      - in: path
        name: coalition_id
        required: true
        schema:
          title: BSON object id
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutCoalitionIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCoalitionOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update a coalition
      tags:
      - Coalition management
  /coalition/{coalition_id}/tracking:
    get:
      description: 'Reports which member (of a given coalition) have a valid refresh
        token

        attacked to them, and which do not. Requires a clearance level of 5 or

        more.'
      operationId: get_coalition_tracking_coalition__coalition_id__tracking_get
      parameters:
      - in: path
        name: coalition_id
        required: true
        schema:
          title: BSON object id
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTrackingOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Coalition tracking
      tags:
      - Coalition management
...